+++ parser/linter_objects.c
@@ -24,7 +24,7 @@ static void	lint_cylinder_extra(char *line, t_scene *scene, int count)
 
 void	lint_cylinder(char *line, t_scene *scene)
 {
-	int	count;
+	size_t	count;
 
 	count = ft_split_inplace(line, ' ');
 	if (count < 6 || count > 12)
@@ -68,7 +68,7 @@ static void	lint_cone_extra(char *line, t_scene *scene, int count)
 
 void	lint_cone(char *line, t_scene *scene)
 {
-	int	count;
+	size_t	count;
 
 	count = ft_split_inplace(line, ' ');
 	if (count < 6 || count > 12)
